pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "one-pos-utilidades"
version = "1.0.0"
description = "Servidor de impresión ESC/POS ligero para retail"
authors = [{name = "I-Labs Chile"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.0",
    "pydantic==2.6.4",
    "python-multipart==0.0.9",
    "Pillow==10.3.0",
    "numpy==1.26.4",
    "pyusb==1.2.1",
    "requests==2.32.3",
    "pywin32; sys_platform == 'win32'",
]

[project.urls]
Homepage = "https://github.com/I-Labs-Chile/ONE-POS-Utilidades"
Repository = "https://github.com/I-Labs-Chile/ONE-POS-Utilidades"
